move jenkins related ext tests to image_ecosystem
